video-dublagem-automatica/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
yt-dlp==2023.11.16
ffmpeg-python==0.2.1
moviepy==1.0.3
google-api-python-client==2.101.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
